# Small improvements (performance, ensure valid constructor usage)