--- a/lib/parser/hiredis.js
+++ b/lib/parser/hiredis.js
@@ -1,3 +1,4 @@
+'use strict';
 var events = require("events"),
     util = require("../util"),
     hiredis = require("hiredis");