--- a/lib/queue.js
+++ b/lib/queue.js
@@ -1,6 +1,8 @@
 // Queue class adapted from Tim Caswell's pattern library
 // http://github.com/creationix/pattern/blob/master/lib/pattern/queue.js
 
+'use strict';
+
 function Queue() {
     this.tail = [];
     this.head = [];
@@ -50,7 +52,8 @@ Queue.prototype.getLength = function () {
 Object.defineProperty(Queue.prototype, "length", {
     get: function () {
         return this.getLength();
-    }
+    },
+    configurable: false
 });