--- a/lib/to_array.js
+++ b/lib/to_array.js
@@ -1,12 +1,6 @@
-function to_array(args) {
-    var len = args.length,
-        arr = new Array(len), i;
-
-    for (i = 0; i < len; i += 1) {
-        arr[i] = args[i];
-    }
-
-    return arr;
-}
-
-module.exports = to_array;
+module.exports = (function(){
+	var slice = [].slice;
+	return function to_array(arr){
+		return slice.call(arr)
+	};
+}());