* Increased perfomance of `to_array` method.
Following good rule "any perfomance should be proven by benchmark" I provide following code: https://gist.github.com/Ginden/4b3ca5bc48634fa9500c
```
toArray1 took 13.677ms on 32768 arrays
toArray2 [current] took 20.659ms on 32768 arrays
toArray3 [new] took 11.943ms on 32768 arrays
toArray4 took 12.269ms on 32768 arrays
```

* `'use strict'` directive was added to enforce valid constructor usage. It's minor "potentially breaking" change - it will make already buggy code fail earlier.