@BridgeAR 2015-07-22
@Ginden Your benchmark does not really test what you think it does. toArray1, toArray3 and toArray4 are all using the same functions internal and the first is going to warm the other ones up. In fact your test does not take that into acount. v8 has a jit and has to be warmed up before comparing. So you should have warmed all functions up. Besides that the dummyObject does not return any value ;).

But that's not really my concern. The to_array function is there for slicing arguments. You did not benchmark this for arguments but for simple arrays.
After all, `[].slice.call(arr);` is indeed faster for simple arrays but the function will leak arguments and deopt in v8 if you use this for arguments.

In that case the slice will be drastically slower than the function in place right now.