@Ginden 2015-07-22
@BridgeAR: 
> Your benchmark does not really test what you think it does. toArray1, toArray3 and toArray4 are all using the same functions internal and the first is going to warm the other ones up.
This statement isn't true. Functions provided in standard library are treated specially, being [almost] always inlined, [almost] always optimized and so on.

>After all, [].slice.call(arr); is indeed faster for simple arrays but the function will leak arguments and deopt in v8 if you use this for arguments.
> In that case the slice will be drastically slower than the function in place right now.
This statement isn't true, too. If you are calling `func(arguments)` deopt does not depend on `func` internals.
BTW - AFAIR it doesn't causes deopt but prevents optimization.

And even if your argument isn't true (`arguments` materialization isn't benchmarked here), your thesis is true - it's slower. https://gist.github.com/Ginden/4b3ca5bc48634fa9500c
>toArray1 took 29.041ms on 32768 arrays
>toArray2 took 11.748ms on 32768 arrays
>toArray3 took 19.089ms on 32768 arrays
> toArray4 took 19.222ms on 32768 arrays


>Besides that the dummyObject does not return any value 
This statement is true.
>You did not benchmark this for arguments but for simple arrays.
You are right.