@mranney 2015-07-22
It wouldn't surprise me if at some point V8 would implement a faster way to do this. But to_array is there because this process kept showing up on the flame graphs.